fix: format add attribution log
