+++ src/app/(proper_react)/(redesign)/(authenticated)/user/(dashboard)/dashboard/page.tsx
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { headers, cookies } from "next/headers";
+import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 import { getServerSession } from "next-auth";
 import { View } from "./View";
@@ -32,10 +32,11 @@ import {
 import { refreshStoredScanResults } from "../../../../../../functions/server/refreshStoredScanResults";
 import { getEnabledFeatureFlags } from "../../../../../../../db/tables/featureFlags";
 import { parseIso8601Datetime } from "../../../../../../../utils/parse";
-import {
-  addAttributionForSubscriber,
-  getLatestAttributionForSubscriberWithType,
-} from "../../../../../../../db/tables/attributions";
+// import {
+//   addAttributionForSubscriber,
+//   getLatestAttributionForSubscriberWithType,
+// } from "../../../../../../../db/tables/attributions";
+import { getAttributionsFromCookiesOrDb } from "../../../../../../functions/server/attributions";
 import { getUserId } from "../../../../../../functions/server/getUserId";
 
 export default async function DashboardPage() {
@@ -45,7 +46,7 @@ export default async function DashboardPage() {
   }
 
   const headersList = headers();
-  const cookiesList = cookies();
+  // const cookiesList = cookies();
   const countryCode = getCountryCode(headersList);
 
   const profileId = await getOnerepProfileId(session.user.subscriber.id);
@@ -117,49 +118,53 @@ export default async function DashboardPage() {
   const yearlySubscriptionUrl = getPremiumSubscriptionUrl({ type: "yearly" });
   const fxaSettingsUrl = process.env.FXA_SETTINGS_URL!;
   const profileStats = await getProfilesStats();
-  let additionalSubplatParams = new URLSearchParams(
-    cookiesList.get("attributionsLastTouch")?.value,
+  const additionalSubplatParams = await getAttributionsFromCookiesOrDb(
+    session.user.subscriber.id,
   );
+  console.log({ additionalSubplatParams });
+  // let additionalSubplatParams = new URLSearchParams(
+  //   cookiesList.get("attributionsLastTouch")?.value,
+  // );
 
-  // store utm attributions if present in cookies
-  if (cookiesList.get("attributionsFirstTouch")?.value) {
-    const searchParams = new URLSearchParams(
-      cookiesList.get("attributionsFirstTouch")?.value,
-    );
-    const attribution = {
-      type: "firstTouch",
-      entrypoint: searchParams.get("entrypoint") ?? "",
-      utm_campaign: searchParams.get("utm_campaign") ?? "",
-      utm_medium: searchParams.get("utm_medium") ?? "",
-      utm_source: searchParams.get("utm_source") ?? "",
-      utm_term: searchParams.get("utm_term") ?? "",
-    };
-    await addAttributionForSubscriber(session.user.subscriber.id, attribution);
-  }
+  // // store utm attributions if present in cookies
+  // if (cookiesList.get("attributionsFirstTouch")?.value) {
+  //   const searchParams = new URLSearchParams(
+  //     cookiesList.get("attributionsFirstTouch")?.value,
+  //   );
+  //   const attribution = {
+  //     type: "firstTouch",
+  //     entrypoint: searchParams.get("entrypoint") ?? "",
+  //     utm_campaign: searchParams.get("utm_campaign") ?? "",
+  //     utm_medium: searchParams.get("utm_medium") ?? "",
+  //     utm_source: searchParams.get("utm_source") ?? "",
+  //     utm_term: searchParams.get("utm_term") ?? "",
+  //   };
+  //   await addAttributionForSubscriber(session.user.subscriber.id, attribution);
+  // }
 
-  if (additionalSubplatParams.size > 0) {
-    const attribution = {
-      type: "lastTouch",
-      entrypoint: additionalSubplatParams.get("entrypoint") ?? "",
-      utm_campaign: additionalSubplatParams.get("utm_campaign") ?? "",
-      utm_medium: additionalSubplatParams.get("utm_medium") ?? "",
-      utm_source: additionalSubplatParams.get("utm_source") ?? "",
-      utm_term: additionalSubplatParams.get("utm_term") ?? "",
-    };
-    await addAttributionForSubscriber(session.user.subscriber.id, attribution);
-  } else {
-    // if "attributionsLastTouch" cookie isn't present, try to load the attribution from the db
-    const attributionLastTouch =
-      await getLatestAttributionForSubscriberWithType(
-        session.user.subscriber.id,
-        "lastTouch",
-      );
-    if (attributionLastTouch) {
-      additionalSubplatParams = new URLSearchParams(
-        attributionLastTouch as Record<string, string>,
-      );
-    }
-  }
+  // if (additionalSubplatParams.size > 0) {
+  //   const attribution = {
+  //     type: "lastTouch",
+  //     entrypoint: additionalSubplatParams.get("entrypoint") ?? "",
+  //     utm_campaign: additionalSubplatParams.get("utm_campaign") ?? "",
+  //     utm_medium: additionalSubplatParams.get("utm_medium") ?? "",
+  //     utm_source: additionalSubplatParams.get("utm_source") ?? "",
+  //     utm_term: additionalSubplatParams.get("utm_term") ?? "",
+  //   };
+  //   await addAttributionForSubscriber(session.user.subscriber.id, attribution);
+  // } else {
+  //   // if "attributionsLastTouch" cookie isn't present, try to load the attribution from the db
+  //   const attributionLastTouch =
+  //     await getLatestAttributionForSubscriberWithType(
+  //       session.user.subscriber.id,
+  //       "lastTouch",
+  //     );
+  //   if (attributionLastTouch) {
+  //     additionalSubplatParams = new URLSearchParams(
+  //       attributionLastTouch as Record<string, string>,
+  //     );
+  //   }
+  // }
 
   return (
     <View
